+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to MiniJinja are documented here.
 
 # 0.15.0
 
-- Removed the functionality to unload templates from an environment. (#60)
 - Bumped minimum version requirement to 1.43.
+- Internal refactorings.
+- Added support for fully dynamic loading via `Source::set_loader`.
 
 # 0.14.1